+++ app/components/templates/index.ts
@@ -7,3 +7,4 @@ export * from './spacer'
 export * from './status-button'
 export * from './textarea-field'
 export * from './toaster'
+export * from './user-dropdown'
