+++ app/root.tsx
@@ -11,7 +11,6 @@ import {
 	type MetaFunction,
 } from '@remix-run/node'
 import {
-	Form,
 	Link,
 	Links,
 	LiveReload,
@@ -22,38 +21,26 @@ import {
 	useFetcher,
 	useFetchers,
 	useLoaderData,
-	useSubmit,
 } from '@remix-run/react'
 import { withSentry } from '@sentry/remix'
-import { useRef } from 'react'
 import { HoneypotProvider } from 'remix-utils/honeypot/react'
 import { z } from 'zod'
 import { GeneralErrorBoundary } from './components/layout'
-import { EpicProgress, useToast } from './components/templates'
-import {
-	Button,
-	DropdownMenu,
-	DropdownMenuContent,
-	DropdownMenuItem,
-	DropdownMenuPortal,
-	DropdownMenuTrigger,
-	Icon,
-	href as iconsHref,
-	EpicToaster,
-} from './components/ui'
+import { EpicProgress, useToast, UserDropdown } from './components/templates'
+import { Icon, href as iconsHref, EpicToaster } from './components/ui'
 import tailwindStyleSheetUrl from './styles/tailwind.css'
 import { getUserId, logout } from './utils/auth.server.ts'
 import { ClientHintCheck, getHints, useHints } from './utils/client-hints.tsx'
 import { prisma } from './utils/db.server.ts'
 import { getEnv } from './utils/env.server.ts'
 import { honeypot } from './utils/honeypot.server.ts'
-import { combineHeaders, getDomainUrl, getUserImgSrc } from './utils/misc.tsx'
+import { combineHeaders, getDomainUrl } from './utils/misc.tsx'
 import { useNonce } from './utils/nonce-provider.ts'
 import { useRequestInfo } from './utils/request-info.ts'
 import { type Theme, setTheme, getTheme } from './utils/theme.server.ts'
 import { makeTimings, time } from './utils/timing.server.ts'
 import { getToast } from './utils/toast.server.ts'
-import { useOptionalUser, useUser } from './utils/user.ts'
+import { useOptionalUser } from './utils/user.ts'
 
 export const links: LinksFunction = () => {
 	return [
@@ -281,66 +268,66 @@ function AppWithProviders() {
 
 export default withSentry(AppWithProviders)
 
-function UserDropdown() {
-	const user = useUser()
-	const submit = useSubmit()
-	const formRef = useRef<HTMLFormElement>(null)
-	return (
-		<DropdownMenu>
-			<DropdownMenuTrigger asChild>
-				<Button asChild variant="secondary">
-					<Link
-						to={`/users/${user.username}`}
-						// this is for progressive enhancement
-						onClick={e => e.preventDefault()}
-						className="flex items-center gap-2"
-					>
-						<img
-							className="h-8 w-8 rounded-full object-cover"
-							alt={user.name ?? user.username}
-							src={getUserImgSrc(user.image?.id)}
-						/>
-						<span className="text-body-sm font-bold">
-							{user.name ?? user.username}
-						</span>
-					</Link>
-				</Button>
-			</DropdownMenuTrigger>
-			<DropdownMenuPortal>
-				<DropdownMenuContent sideOffset={8} align="start">
-					<DropdownMenuItem asChild>
-						<Link prefetch="intent" to={`/users/${user.username}`}>
-							<Icon className="text-body-md" name="avatar">
-								Profile
-							</Icon>
-						</Link>
-					</DropdownMenuItem>
-					<DropdownMenuItem asChild>
-						<Link prefetch="intent" to={`/users/${user.username}/notes`}>
-							<Icon className="text-body-md" name="pencil-2">
-								Notes
-							</Icon>
-						</Link>
-					</DropdownMenuItem>
-					<DropdownMenuItem
-						asChild
-						// this prevents the menu from closing before the form submission is completed
-						onSelect={event => {
-							event.preventDefault()
-							submit(formRef.current)
-						}}
-					>
-						<Form action="/logout" method="POST" ref={formRef}>
-							<Icon className="text-body-md" name="exit">
-								<button type="submit">Logout</button>
-							</Icon>
-						</Form>
-					</DropdownMenuItem>
-				</DropdownMenuContent>
-			</DropdownMenuPortal>
-		</DropdownMenu>
-	)
-}
+// function UserDropdown() {
+// 	const user = useUser()
+// 	const submit = useSubmit()
+// 	const formRef = useRef<HTMLFormElement>(null)
+// 	return (
+// 		<DropdownMenu>
+// 			<DropdownMenuTrigger asChild>
+// 				<Button asChild variant="secondary">
+// 					<Link
+// 						to={`/users/${user.username}`}
+// 						// this is for progressive enhancement
+// 						onClick={e => e.preventDefault()}
+// 						className="flex items-center gap-2"
+// 					>
+// 						<img
+// 							className="h-8 w-8 rounded-full object-cover"
+// 							alt={user.name ?? user.username}
+// 							src={getUserImgSrc(user.image?.id)}
+// 						/>
+// 						<span className="text-body-sm font-bold">
+// 							{user.name ?? user.username}
+// 						</span>
+// 					</Link>
+// 				</Button>
+// 			</DropdownMenuTrigger>
+// 			<DropdownMenuPortal>
+// 				<DropdownMenuContent sideOffset={8} align="start">
+// 					<DropdownMenuItem asChild>
+// 						<Link prefetch="intent" to={`/users/${user.username}`}>
+// 							<Icon className="text-body-md" name="avatar">
+// 								Profile
+// 							</Icon>
+// 						</Link>
+// 					</DropdownMenuItem>
+// 					<DropdownMenuItem asChild>
+// 						<Link prefetch="intent" to={`/users/${user.username}/notes`}>
+// 							<Icon className="text-body-md" name="pencil-2">
+// 								Notes
+// 							</Icon>
+// 						</Link>
+// 					</DropdownMenuItem>
+// 					<DropdownMenuItem
+// 						asChild
+// 						// this prevents the menu from closing before the form submission is completed
+// 						onSelect={event => {
+// 							event.preventDefault()
+// 							submit(formRef.current)
+// 						}}
+// 					>
+// 						<Form action="/logout" method="POST" ref={formRef}>
+// 							<Icon className="text-body-md" name="exit">
+// 								<button type="submit">Logout</button>
+// 							</Icon>
+// 						</Form>
+// 					</DropdownMenuItem>
+// 				</DropdownMenuContent>
+// 			</DropdownMenuPortal>
+// 		</DropdownMenu>
+// 	)
+// }
 
 /**
  * @returns the user's theme preference, or the client hint theme if the user
